Extract separate factory
